pyproject.toml
[project]
name = "simulacra"
version = "2.34.0"
requires-python = "~=3.12.0"
dependencies = [
    "openai",
    "pytz",
    "pyyaml",
    "python-telegram-bot",
    "ruamel.yaml",
    "toml",
    "jinja2",
    "httpx",
    "pdfplumber",
    "backoff",
    "rapidfuzz",
    "dotenv",
]

[dependency-groups]
dev = [
    "ruff",
    "hupper",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pyfakefs",
    "pytest-httpx",
    "pytest-cov",
    "mypy",
    "types-pyyaml",
    "types-pytz",
    "types-toml",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]

# https://github.com/pytest-dev/pytest-asyncio/issues/924
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
